clean up HRID handling
