+++ src/folio_migration_tools/migration_tasks/holdings_csv_transformer.py
@@ -29,6 +29,7 @@ from folio_migration_tools.mapping_file_transformation.holdings_mapper import (
 from folio_migration_tools.mapping_file_transformation.mapping_file_mapper_base import (
     MappingFileMapperBase,
 )
+from folio_migration_tools.marc_rules_transformation.hrid_handler import HRIDHandler
 from folio_migration_tools.migration_tasks.migration_task_base import MigrationTaskBase
 from folio_migration_tools.task_configuration import AbstractTaskConfiguration
 
@@ -130,7 +131,10 @@ class HoldingsCsvTransformer(MigrationTaskBase):
                 self.task_configuration.reset_hrid_settings
                 and not self.task_configuration.never_update_hrid_settings
             ):
-                self.mapper.reset_holdings_hrid_counter()
+                hrid_handler = HRIDHandler(
+                    self.folio_client, HridHandling.default, self.mapper.migration_report, True
+                )
+                hrid_handler.reset_holdings_hrid_counter()
 
         except HTTPError as http_error:
             logging.critical(http_error)
