--- tests/test_data/diacritics/diac_from_oclc.mrc:Zone.Identifier
@@ -1,2 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
