+++ tests/test_holdings_processor.py
@@ -3,12 +3,16 @@ from unittest.mock import Mock
 from pymarc import Field
 from pymarc import Record
 
+from folio_migration_tools.library_configuration import HridHandling
+from folio_migration_tools.marc_rules_transformation.hrid_handler import HRIDHandler
 from folio_migration_tools.marc_rules_transformation.marc_file_processor import (
     MarcFileProcessor,
 )
 from folio_migration_tools.marc_rules_transformation.rules_mapper_holdings import (
     RulesMapperHoldings,
 )
+from folio_migration_tools.migration_report import MigrationReport
+from folio_migration_tools.test_infrastructure import mocked_classes
 
 
 def test_chop_008():
@@ -25,16 +29,24 @@ def test_chop_008():
 def test_generate_num_part_retain_leading_zeroes():
     mock_processor = Mock(spec=MarcFileProcessor)
     mock_mapper = Mock(spec=RulesMapperHoldings)
-    mock_mapper.holdings_hrid_counter = "1"
-    mock_mapper.common_retain_leading_zeroes = True
+
     mock_processor.mapper = mock_mapper
-    assert MarcFileProcessor.generate_num_part(mock_processor) == "00000000001"
+    hrid_handler = HRIDHandler(
+        mocked_classes.mocked_folio_client(), HridHandling.preserve001, MigrationReport(), True
+    )
+    hrid_handler.common_retain_leading_zeroes = True
+    hrid_handler.holdings_hrid_counter = "1"
+    assert hrid_handler.generate_numeric_part(hrid_handler.holdings_hrid_counter) == "00000000001"
 
 
-def test_generate_num_part_without_leading_zeroes():
+def test_generate_num_part_no_retain_leading_zeroes():
     mock_processor = Mock(spec=MarcFileProcessor)
     mock_mapper = Mock(spec=RulesMapperHoldings)
-    mock_mapper.holdings_hrid_counter = "1"
-    mock_mapper.common_retain_leading_zeroes = False
+
     mock_processor.mapper = mock_mapper
-    assert MarcFileProcessor.generate_num_part(mock_processor) == "1"
+    hrid_handler = HRIDHandler(
+        mocked_classes.mocked_folio_client(), HridHandling.preserve001, MigrationReport(), True
+    )
+    hrid_handler.common_retain_leading_zeroes = False
+    hrid_handler.holdings_hrid_counter = "1"
+    assert hrid_handler.generate_numeric_part(hrid_handler.holdings_hrid_counter) == "1"
